refactor: move scene modules to core/scene
