+++ src/App.tsx
@@ -5,7 +5,7 @@ import TopBar from './components/TopBar'
 import Sidebar from './components/Sidebar'
 import StatsWidget from './components/StatsWidget'
 import CameraHints from './components/CameraHints'
-import Viewport from './scene/Viewport'
+import Viewport from './core/scene/Viewport'
 import { useStore } from './store/useStore'
 import { encodePieces } from './core/utils/serialization'
 import { usePersistence } from './hooks/usePersistence'
