+++ src/core/scene/HitPlane.tsx
@@ -1,17 +1,17 @@
 import { useState } from 'react'
 import type { ThreeEvent } from '@react-three/fiber'
-import { useStore } from '../store/useStore'
-import { canPlace, canPlaceTriSnap, canPlaceTriSnapEdge, canPlaceSquareSnap, canPlaceSquareSnapEdge } from '../utils/validation'
+import { useStore } from '../../store/useStore'
+import { canPlace, canPlaceTriSnap, canPlaceTriSnapEdge, canPlaceSquareSnap, canPlaceSquareSnapEdge } from '../../utils/validation'
 import { detectSide, getCellPieceShape } from './pieceGeometry'
-import piecesConfig from '../data/pieces-config.json'
-import type { PiecesConfig, XYZ, PieceSide, PieceRotation, PlacedPiece } from '../core/types'
+import piecesConfig from '../../data/pieces-config.json'
+import type { PiecesConfig, XYZ, PieceSide, PieceRotation, PlacedPiece } from '../types'
 import GhostPiece from './GhostPiece'
-import { worldToTriCoord, triSlotWorldPosition, triSlotRotationDeg, triEdgeWorldPosition, triEdgeRotationDeg, detectTriEdge, squareEdgeSnapPosition, triSnapNeighbors, triSnapEdgeWorldPosition, triSnapEdgeRotationDeg, detectTriSnapEdge, triEdgeSquareSnapPosition, triSnapEdgeSquareSnapPosition, squareSnapEdgeWorldPosition, squareSnapEdgeRotationDeg, detectSquareSnapSide, HEX_ORIGIN } from '../core/utils/hexGrid'
-import { toKey, toEdgeKey, toTriKey, toTriEdgeKey, toTriSnapKey, toSquareSnapKey } from '../core/utils/coordinateKey'
+import { worldToTriCoord, triSlotWorldPosition, triSlotRotationDeg, triEdgeWorldPosition, triEdgeRotationDeg, detectTriEdge, squareEdgeSnapPosition, triSnapNeighbors, triSnapEdgeWorldPosition, triSnapEdgeRotationDeg, detectTriSnapEdge, triEdgeSquareSnapPosition, triSnapEdgeSquareSnapPosition, squareSnapEdgeWorldPosition, squareSnapEdgeRotationDeg, detectSquareSnapSide, HEX_ORIGIN } from '../utils/hexGrid'
+import { toKey, toEdgeKey, toTriKey, toTriEdgeKey, toTriSnapKey, toSquareSnapKey } from '../utils/coordinateKey'
 import { GHOST_VALID_COLOR } from './pieceGeometry'
 import CellMesh from './CellMesh'
 import EdgeMesh from './EdgeMesh'
-import type { TriCoord, TriEdgeIndex, TriSnapTarget, SquareSnapTarget } from '../core/types'
+import type { TriCoord, TriEdgeIndex, TriSnapTarget, SquareSnapTarget } from '../types'
 
 const config = piecesConfig as PiecesConfig
 const GRID_W = 5
