+++ src/core/scene/woodTexture.ts
